Moved udfgen2udf converters to the node. The controller/node imports were mixed breaking the docker images creation.
